Refactor which methods are controller or helper, move error definition
