--- app/lib/attachment_not_found_error.rb
@@ -1,1 +1,0 @@
-class AttachmentNotFoundError < StandardError; end
